requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
Werkzeug
SQLAlchemy
Flask-Migrate
python-dotenv
